+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Don't emit empty `interfaces` in type hints.
 - Don't `try/except` imports in `.pyi` files.
 - Don't use `typing.Optional` in generic type parameters.
+- Fully implement mapping protocol for `IMap`/`IMapView`.
 
 ### Fixed
 - Fixed support for sub-interpreters.